Added participant generator for demo purposes

+++ participant-tracking/src/main/java/org/ikuven/bbut/tracking/ResultTrackingApplication.java
@@ -1,16 +1,16 @@
 package org.ikuven.bbut.tracking;
 
-import org.ikuven.bbut.tracking.participant.Lap;
-import org.ikuven.bbut.tracking.participant.LapState;
+import org.ikuven.bbut.tracking.generation.ParticipantGenerator;
 import org.ikuven.bbut.tracking.repository.DatabaseSequenceRepository;
 import org.ikuven.bbut.tracking.repository.ParticipantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.core.env.Environment;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
-import java.time.LocalDateTime;
+import java.util.Arrays;
 
 @SpringBootApplication
 @EnableMongoRepositories
@@ -26,24 +26,22 @@ public class ResultTrackingApplication implements CommandLineRunner {
     @Autowired
     DatabaseSequenceRepository sequenceRepository;
 
+    @Autowired
+    ParticipantGenerator participantGenerator;
+
+    @Autowired
+    Environment environment;
+
     @Override
     public void run(String... args) {
 
-        sequenceRepository.deleteAll();
-        repository.deleteAll();
-//
-//        Participant participant = Participant.of(1, "Ken", "Alexandersson", Collections.singletonList(createLap(1, LapState.NOT_STARTED)));
-//        Participant participant2 = Participant.of(2, "Torbjörn", "Grahn", Arrays.asList(createLap(1, LapState.COMPLETED), createLap(2, LapState.COMPLETED)));
-//        Participant participant3 = Participant.of(3, "Sune", "Bom", Arrays.asList(createLap(1, LapState.COMPLETED), createLap(2, LapState.STARTED)));
-//        Participant participant4 = Participant.of(4, "Vera", "Blut", Collections.emptyList());
-//
-////        repository.save(participant);
-////        repository.save(participant2);
-////        repository.save(participant3);
-//        repository.save(participant4);
-    }
+        if (Arrays.asList(environment.getActiveProfiles()).contains("dev")) {
+            sequenceRepository.deleteAll();
+            repository.deleteAll();
+        }
 
-    private Lap createLap(int lapNumber, LapState lapState) {
-        return Lap.of(lapNumber, LocalDateTime.now(), lapState);
+        if (Arrays.asList(environment.getActiveProfiles()).contains("demo")) {
+            participantGenerator.generate();
+        }
     }
 }
